Add modules `utils/backend.py` for backend imports (moved from __init__.py) and `utils/deprecation.py` for deprecation util functions (moved from __init__.py).
